feat(mobile): init design system
