+++ apps/mobile/index.js
@@ -1,8 +1,8 @@
 import { registerRootComponent } from 'expo';
 
-import App from './App';
+import Main from './src/main';
 
 // registerRootComponent calls AppRegistry.registerComponent('main', () => App);
 // It also ensures that whether you load the app in the Expo client or in a native build,
 // the environment is set up appropriately
-registerRootComponent(App);
+registerRootComponent(Main);
