+++ apps/mobile/package.json
@@ -9,12 +9,16 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@judgeme/react": "workspace:*",
     "expo": "~47.0.9",
+    "expo-navigation-bar": "~2.0.1",
     "expo-status-bar": "~1.4.2",
     "expo-updates": "^0.15.6",
     "react": "18.1.0",
     "react-dom": "18.1.0",
-    "react-native": "0.70.5",
+    "react-native": "0.70.6",
+    "react-native-paper": "^5.1.2",
+    "react-native-safe-area-context": "^4.4.1",
     "react-native-web": "~0.18.9"
   },
   "devDependencies": {
