+++ apps/mobile/src/main.tsx
@@ -1,0 +1,11 @@
+import { UiProvider } from '@judgeme/react';
+
+import App from './App';
+
+export default function Main() {
+  return (
+    <UiProvider>
+      <App />
+    </UiProvider>
+  );
+}
